requirements.txt
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cssselect==1.1.0
dateparser==1.0.0
demjson==2.2.4
Django==3.2.4
django-crispy-forms==1.12.0
django-filter==2.4.0
django-lockdown==4.0.0
facebook-scraper==0.2.43
fake-useragent==0.1.11
google-api-core==1.30.0
google-auth==1.32.0
google-cloud==0.34.0
google-cloud-vision==2.3.2
googleapis-common-protos==1.53.0
grpcio==1.38.1
idna==2.10
install==1.3.4
lxml==4.6.3
packaging==20.9
parse==1.19.0
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.17.3

pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.1.0
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-html==0.10.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tqdm==4.61.1
tzlocal==2.1
urllib3==1.26.5
w3lib==1.22.0
websockets==8.1
